requirements.txt
python-telegram-bot==20.7
openai==1.12.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
pydantic==2.5.3
redis==5.0.1
pydub==0.25.1